# lib/api-zod/src/generated/types/printerInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Nativos 3D pricing and production records
 * OpenAPI spec version: 0.1.0
 */

export interface PrinterInput {
  /** @minLength 1 */
  name: string;
  model: string;
  /** @minimum 0 */
  purchaseValue: number;
  /** @minimum 0 */
  residualValue: number;
  /** @minimum 0 */
  usefulLifeHours: number;
  /** @minimum 0 */
  averagePowerWatts: number;
  /** @minimum 0 */
  maintenancePerHour: number;
  active: boolean;
}